+++ lib/art.js
@@ -137,6 +137,20 @@ exports.poweroff = ({phase = 'pre'} = {}) => {
 /*
  * Helper to show status of secret toggle
  */
+exports.print = ({text, color = 'white'} = {}) => {
+  return chalk[color](text);
+};
+
+/*
+ * Helper to show status of secret toggle
+ */
+exports.printFont = ({text, color = 'magenta', font = 'Small Slant'} = {}) => {
+  return chalk[color](niceFont(text, 'Small Slant'));
+};
+
+/*
+ * Helper to show status of secret toggle
+ */
 exports.secretToggle = (on = false) => {
   switch (on) {
     case true:
