+++ lib/formatters.js
@@ -59,6 +59,7 @@ exports.formatData = (data, {path = '', format = 'default', filter = []} = {}, o
         colors: process.stdout.isTTY,
         depth: 10,
         compact: true,
+        sorted: _.get(opts, 'sort', false),
       });
   }
 };
