+++ lib/table.js
@@ -29,7 +29,7 @@ const noBorderOpts = {
 };
 
 module.exports = class Table extends OldTable {
-  constructor(data, {border = true, keyColor = 'cyan', joiner = os.EOL, sort = true} = {}, opts = {}) {
+  constructor(data, {border = true, keyColor = 'cyan', joiner = os.EOL, sort = false} = {}, opts = {}) {
     // Inherit the table
     const tableDefaults = border ? {} : noBorderOpts;
     super(_.merge({}, tableDefaults, opts));
