+++ plugins/lando-core/tasks/rebuild.js
@@ -41,7 +41,7 @@ module.exports = lando => {
         console.log(lando.cli.makeArt('appRebuild', {name: app.name, phase: 'pre'}));
         return app.rebuild().then(() => {
           console.log(lando.cli.makeArt('appRebuild', {name: app.name, phase: 'post'}));
-          console.log(lando.cli.formatData(utils.startTable(app), {format: 'table'}, {sort: false, border: false}));
+          console.log(lando.cli.formatData(utils.startTable(app), {format: 'table'}, {border: false}));
           console.log('');
         });
       }
