+++ plugins/lando-core/tasks/start.js
@@ -14,7 +14,7 @@ module.exports = lando => {
         console.log(lando.cli.makeArt('appStart', {name: app.name, phase: 'pre'}));
         return app.start().then(() => {
           console.log(lando.cli.makeArt('appStart', {name: app.name, phase: 'post'}));
-          console.log(lando.cli.formatData(utils.startTable(app), {format: 'table'}, {sort: false, border: false}));
+          console.log(lando.cli.formatData(utils.startTable(app), {format: 'table'}, {border: false}));
           console.log('');
         });
       }
